prov/sm2: remove sm2_util.c
Combine sm2_util.c with sm2_init because sm2_util.c only contained
initializtion code.
