+++ prov/sm2/Makefile.include
@@ -12,7 +12,6 @@ _sm2_files = \
 	prov/sm2/src/sm2_init.c		\
 	prov/sm2/src/sm2_av.c		\
 	prov/sm2/src/sm2.h 		\
-	prov/sm2/src/sm2_util.c		\
 	prov/sm2/src/sm2_atom.h 	\
 	prov/sm2/src/sm2_coordination.c \
 	prov/sm2/src/sm2_coordination.h \
